Commit work in progress
Abandoning this for now, can't figure out how to combine a field that is
derived from ListStringItem with image formatters (I think having a
dedicated avatar formatter might be the wrong route; ideally it should
be possible to select regular Image formatters). Might need to do some
altering for that, somehow.